+++ src/routes/project.ts
@@ -1,5 +1,9 @@
+import { requireAuthenticationMiddleware } from '@/middlewares/authentication.js';
 import { ProjectRepository } from '@/repositories/ProjectRepositories.js';
-import { UpdateProjectSchema } from '@/types/validators/Project.js';
+import {
+  CreateProjectSchema,
+  UpdateProjectSchema,
+} from '@/types/validators/Project.js';
 import { LogDomain, logger } from '@/utils/logger.js';
 import { failure } from '@/views/json/failure.js';
 import { success } from '@/views/json/success.js';
@@ -7,9 +11,13 @@ import { Hono } from 'hono';
 
 const project = new Hono();
 
-project.get('/', async (_c) => {
+project.use(requireAuthenticationMiddleware);
+
+project.get('/', async (c) => {
   try {
-    const result = await ProjectRepository.readAll();
+    logger.info([LogDomain.ROUTE], 'Reading all projects');
+    const user = c.get('user');
+    const result = await ProjectRepository.readAll({ userId: user.id });
     return success(200, result);
   } catch (err) {
     logger.error([LogDomain.ROUTE], 'Error when reading projects', { error: err });
@@ -19,8 +27,15 @@ project.get('/', async (_c) => {
 
 project.get('/:id', async (c) => {
   try {
+    logger.info([LogDomain.ROUTE], 'Reading project', {
+      projectId: c.req.param('id'),
+    });
     const id = c.req.param('id');
-    const result = await ProjectRepository.readOne(id);
+    const user = c.get('user');
+    const result = await ProjectRepository.readOne({
+      projectId: id,
+      userId: user.id,
+    });
     return success(200, result);
   } catch (err) {
     return failure(err);
@@ -29,9 +44,12 @@ project.get('/:id', async (c) => {
 
 project.post('/', async (c) => {
   try {
+    logger.info([LogDomain.ROUTE], 'Creating project');
+    const user = c.get('user');
     const body = await c.req.json();
-    const project = ProjectRepository.create(body);
-    return success(201, project);
+    const project = CreateProjectSchema.parse({ ...body, userId: user.id });
+    const result = ProjectRepository.create(project);
+    return success(201, result);
   } catch (err) {
     return failure(err);
   }
@@ -39,10 +57,18 @@ project.post('/', async (c) => {
 
 project.put('/:id', async (c) => {
   try {
+    logger.info([LogDomain.ROUTE], 'Updating project', {
+      projectId: c.req.param('id'),
+    });
     const id = c.req.param('id');
+    const user = c.get('user');
     const body = await c.req.json();
     const changes = UpdateProjectSchema.parse(body);
-    const result = ProjectRepository.update(id, changes);
+    const result = ProjectRepository.update({
+      projectId: id,
+      userId: user.id,
+      changes: changes,
+    });
     return success(200, result);
   } catch (err) {
     return failure(err);
@@ -51,8 +77,12 @@ project.put('/:id', async (c) => {
 
 project.delete('/:id', async (c) => {
   try {
+    logger.info([LogDomain.ROUTE], 'Deleting project', {
+      projectId: c.req.param('id'),
+    });
     const id = c.req.param('id');
-    const result = await ProjectRepository.delete(id);
+    const user = c.get('user');
+    const result = await ProjectRepository.delete({ projectId: id, userId: user.id });
     return success(200, result);
   } catch (err) {
     return failure(err);
